Bringing supported IDEs to the newest setup, add device and others.
git-svn-id: file:///fltk/svn/fltk/branches/branch-1.3@7345 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ src/fl_ask.cxx
@@ -67,7 +67,7 @@ static char avoidRecursion = 0;
 
 // sets the global return value (ret_val) and closes the window
 static void button_cb(Fl_Widget *bt, void *val) {
-  ret_val = (int)val;
+  ret_val = (int)(long)val;
   message_form->hide();
 }
